LayoutTests/platform/chromium/fast/js/function-names-expected.txt

This test checks the names of all sorts of different functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Function(' return 1; ').toString().replace(/[ \n]+/g, ' ') is 'function anonymous() { return 1; }'
PASS document.documentElement.onclick.toString().replace(/[ \n]+/g, ' ') is 'function onclick(event) { return 2; }'
PASS ''.constructor is String
FAIL Boolean.toString() should be function Boolean() {
    [native code]
}. Was function Boolean() { [native code] }.
FAIL Date.toString() should be function Date() {
    [native code]
}. Was function Date() { [native code] }.
FAIL Error.toString() should be function Error() {
    [native code]
}. Was function Error() { [native code] }.
FAIL EvalError.toString() should be function EvalError() {
    [native code]
}. Was function EvalError() { [native code] }.
FAIL Function.toString() should be function Function() {
    [native code]
}. Was function Function() { [native code] }.
FAIL Number.toString() should be function Number() {
    [native code]
}. Was function Number() { [native code] }.
FAIL Object.toString() should be function Object() {
    [native code]
}. Was function Object() { [native code] }.
FAIL RangeError.toString() should be function RangeError() {
    [native code]
}. Was function RangeError() { [native code] }.
FAIL ReferenceError.toString() should be function ReferenceError() {
    [native code]
}. Was function ReferenceError() { [native code] }.
FAIL RegExp.toString() should be function RegExp() {
    [native code]
}. Was function RegExp() { [native code] }.
FAIL String.toString() should be function String() {
    [native code]
}. Was function String() { [native code] }.
FAIL SyntaxError.toString() should be function SyntaxError() {
    [native code]
}. Was function SyntaxError() { [native code] }.
FAIL TypeError.toString() should be function TypeError() {
    [native code]
}. Was function TypeError() { [native code] }.
FAIL URIError.toString() should be function URIError() {
    [native code]
}. Was function URIError() { [native code] }.
PASS successfullyParsed is true

TEST COMPLETE
